Style: 주석 추가 및 주석 타이틀에 밑줄 추가
+) Spring Security OAuth2 관련 class path 변경
