+++ server/src/main/java/com/soothee/member/repository/MemberRepository.java
@@ -10,11 +10,18 @@ import java.util.Optional;
 @Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
     /**
-     * 회원 닉네임으로 회원 조회
+     * 회원 닉네임으로 회원 조회</hr>
+     *
      * @param memberName String : 조회할 회원 닉네임
      * @return Member : 해당 회원
      */
-    Member findByMemberName(String memberName);
+    Optional<Member> findByMemberName(String memberName);
 
+    /** 인증 회원 식별자와 SNS 종류로 회원 조회</hr>
+     *
+     * @param oauth2ClientId String : 인증 회원 식별자
+     * @param snsType SnsType : KAKAOTALK or GOOGLE
+     * @return Optional<Member> : 조회한 회원(Null 가능)
+     */
     Optional<Member> findByOauth2ClientIdAndSnsType(String oauth2ClientId, SnsType snsType);
 }
