+++ server/src/main/java/com/soothee/oauth2/domain/AuthenticatedUser.java
@@ -1,4 +1,4 @@
-package com.soothee.common.domain;
+package com.soothee.oauth2.domain;
 
 import com.soothee.member.domain.Member;
 import lombok.AllArgsConstructor;
@@ -12,23 +12,29 @@ import org.springframework.security.oauth2.core.user.OAuth2User;
 import javax.security.auth.Subject;
 import java.security.Principal;
 import java.util.Collection;
-import java.util.List;
 import java.util.Map;
+import java.util.stream.Stream;
 
+/**
+ * 인증된 사용자 정보를 시큐리티 컨텍스트(security context)에 보관할 때 사용
+ */
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
 public class AuthenticatedUser implements Principal, OAuth2User {
-    /** 회원 일련번호 */
+    /** OAuth2 인증 회원 식별자 */
     private String oauth2Id;
     /** 회원 이메일 */
     private String email;
-    private List<String> roles;
+    /** 회원 권한 */
+    private String role;
     /** 회원 닉네임 */
     private String memberName;
+    /** 회원 정보 */
     private Map<String, Object> attributes;
 
+    /** 인증된 회원 식별자 가져오기 */
     @Override
     public String getName() {
         return this.oauth2Id;
@@ -46,15 +52,19 @@ public class AuthenticatedUser implements Principal, OAuth2User {
 
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
-        return roles.stream()
-                    .map(SimpleGrantedAuthority::new)
-                    .toList();
+        return Stream.of(role).map(SimpleGrantedAuthority::new).toList();
     }
 
+    /** Member Entity 정보와 OAuth2User 정보로 AuthenticatedUser 생성</hr>
+     *
+     * @param member Member : 서버에 저장된 회원 정보
+     * @param oauth2User OAuth2User : 인증된 회원 정보
+     * @return AuthenticatedUser 회원 정보
+     */
     public static AuthenticatedUser of(Member member, OAuth2User oauth2User) {
         return AuthenticatedUser.builder()
-                                .oauth2Id(String.valueOf(member.getOauth2ClientId()))
-                                .roles(List.of(member.getRole().getRole()))
+                                .oauth2Id(member.getOauth2ClientId())
+                                .role(member.getRole().toString())
                                 .email(member.getEmail())
                                 .memberName(member.getMemberName())
                                 .attributes(oauth2User.getAttributes())
