+++ server/src/main/java/com/soothee/oauth2/domain/GoogleUser.java
@@ -1,4 +1,4 @@
-package com.soothee.common.domain;
+package com.soothee.oauth2.domain;
 
 import com.soothee.common.constants.Role;
 import com.soothee.common.constants.SnsType;
@@ -12,6 +12,10 @@ import java.util.Map;
 public class GoogleUser {
     private final OAuth2User oauth2User;
 
+    /** 인증된 회원 정보를 받아서 Member 생성</hr>
+     *
+     * @return Member : 회원 엔티티
+     */
     public Member toMember() {
         return Member.builder()
                     .email(this.getEmail())
@@ -22,14 +26,17 @@ public class GoogleUser {
                     .build();
     }
 
+    /** 인증 회원 계정 정보 가져오기 */
     private Map<String, Object> getAttributes() {
         return oauth2User.getAttributes();
     }
 
+    /** 인증 회원 정보 중 닉네임 가져오기 */
     private String getNickName() {
         return String.valueOf(this.getAttributes().get("name"));
     }
 
+    /** 인증 회원 정보 중 아이디(이메일) 가져오기 */
     private String getEmail() {
         return String.valueOf(this.getAttributes().get("email"));
     }
