+++ server/src/main/java/com/soothee/oauth2/service/GoogleOAuth2UserService.java
@@ -1,9 +1,8 @@
-package com.soothee.common.service;
+package com.soothee.oauth2.service;
 
 import com.soothee.common.constants.SnsType;
-import com.soothee.common.domain.AuthenticatedUser;
-import com.soothee.common.domain.GoogleUser;
-import com.soothee.common.domain.KakaoUser;
+import com.soothee.oauth2.domain.AuthenticatedUser;
+import com.soothee.oauth2.domain.GoogleUser;
 import com.soothee.member.domain.Member;
 import com.soothee.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
@@ -19,15 +18,26 @@ import java.util.Optional;
 @Transactional
 @RequiredArgsConstructor
 public class GoogleOAuth2UserService implements CustomOAuth2UserService {
-    private static final String REGISTRATION_ID = SnsType.GOOGLE.getSnsType();
+    private static final String REGISTRATION_ID = SnsType.GOOGLE.toString();
     private final MemberService memberService;
 
+    /** 해당 SNS OAuth2을 지원하는지 확인</hr>
+     * "google"인 경우에만 true
+     * @param request
+     * @return
+     */
     @Override
     public boolean supports(OAuth2UserRequest request) {
         String registrationId = request.getClientRegistration().getRegistrationId();
         return StringUtils.equals(REGISTRATION_ID, registrationId);
     }
 
+    /** 인증된 회원 엔티티 조회 및 생성</hr>
+     * 기존 회원 -> 조회
+     * 신규 회원 -> 생성
+     * @param authenticatedUser OAuth2User : 인증된 회원
+     * @return AuthenticatedUser : 인증된 회원 정보를 담은 토큰 반환
+     */
     @Override
     public AuthenticatedUser createOrLoadUser(OAuth2User authenticatedUser) {
         String oauth2ClientId = authenticatedUser.getName();
